Refactor in buffer , gym_manipulator, learner and actor
